question_generation/pyproject.toml
[project]
name = "question-generation"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "dotenv>=0.9.9",
    "fastapi>=0.122.0",
    "langchain>=1.1.0",
    "langchain-openai>=1.1.0",
    "loguru>=0.7.3",
    "psycopg2-binary>=2.9.9",
    "redis>=7.1.0",
]
